+++ value-traits/tests/common.rs
@@ -15,27 +15,27 @@ use value_traits::{
 
 pub fn generic_get<S>(s: S, expected: &[i32])
 where
-    S: SliceByValue<Value = i32> + SliceByValueGet,
+    S: SliceByValueCore<Value = i32> + SliceByValue,
 {
-    assert_eq!(SliceByValue::len(&s), expected.len());
+    assert_eq!(SliceByValueCore::len(&s), expected.len());
 
     for i in 0..expected.len() {
-        assert_eq!(SliceByValueGet::get_value(&s, i), Some(expected[i]));
-        assert_eq!(SliceByValueGet::get_value(&s, i + expected.len()), None);
+        assert_eq!(SliceByValue::get_value(&s, i), Some(expected[i]));
+        assert_eq!(SliceByValue::get_value(&s, i + expected.len()), None);
 
-        assert_eq!(SliceByValueGet::index_value(&s, i), expected[i]);
+        assert_eq!(SliceByValue::index_value(&s, i), expected[i]);
 
         unsafe {
-            assert_eq!(SliceByValueGet::get_value_unchecked(&s, i), expected[i]);
+            assert_eq!(SliceByValue::get_value_unchecked(&s, i), expected[i]);
         }
     }
 }
 
 pub fn generic_slice<S>(s: S, expected: &[i32])
 where
-    S: SliceByValue<Value = i32> + SliceByValueGet + SliceByValueSubslice,
+    S: SliceByValueCore<Value = i32> + SliceByValue + SliceByValueSubslice,
 {
-    assert_eq!(SliceByValue::len(&s), expected.len());
+    assert_eq!(SliceByValueCore::len(&s), expected.len());
 
     let r = ..;
     generic_get(s.index_subslice(r), expected);
@@ -79,31 +79,31 @@ where
 
 pub fn generic_mut<S>(mut s: S)
 where
-    S: SliceByValue<Value = i32> + SliceByValueGet + SliceByValueSet + SliceByValueRepl,
+    S: SliceByValueCore<Value = i32> + SliceByValue + SliceByValueMut,
 {
     for i in 0..s.len() {
-        let old_value = SliceByValueGet::index_value(&s, i);
+        let old_value = SliceByValue::index_value(&s, i);
         let new_value = old_value + 1;
 
         // Test set_value
-        SliceByValueSet::set_value(&mut s, i, new_value);
-        assert_eq!(SliceByValueGet::index_value(&s, i), new_value);
+        SliceByValueMut::set_value(&mut s, i, new_value);
+        assert_eq!(SliceByValue::index_value(&s, i), new_value);
 
         // Test replace_value
-        let replaced_value = SliceByValueRepl::replace_value(&mut s, i, old_value);
+        let replaced_value = SliceByValueMut::replace_value(&mut s, i, old_value);
         assert_eq!(replaced_value, new_value);
-        assert_eq!(SliceByValueGet::index_value(&s, i), old_value);
+        assert_eq!(SliceByValue::index_value(&s, i), old_value);
 
         // test replace_value_unchecked
         let replaced_value =
-            unsafe { SliceByValueRepl::replace_value_unchecked(&mut s, i, new_value) };
+            unsafe { SliceByValueMut::replace_value_unchecked(&mut s, i, new_value) };
         assert_eq!(replaced_value, old_value);
-        assert_eq!(SliceByValueGet::index_value(&s, i), new_value);
+        assert_eq!(SliceByValue::index_value(&s, i), new_value);
 
         // Test unsafe set_value_unchecked
         unsafe {
-            SliceByValueSet::set_value_unchecked(&mut s, i, new_value);
-            assert_eq!(SliceByValueGet::index_value(&s, i), new_value);
+            SliceByValueMut::set_value_unchecked(&mut s, i, new_value);
+            assert_eq!(SliceByValue::index_value(&s, i), new_value);
         }
 
         s.set_value(i, old_value);
@@ -112,13 +112,12 @@ where
 
 pub fn generic_slice_mut<S>(mut s: S)
 where
-    S: SliceByValue<Value = i32>
-        + SliceByValueGet
-        + SliceByValueSet
-        + SliceByValueRepl
+    S: SliceByValueCore<Value = i32>
+        + SliceByValue
+        + SliceByValueMut
         + SliceByValueSubsliceMut,
     for<'a> <S as SliceByValueSubsliceGatMut<'a>>::SubsliceMut:
-        SliceByValue<Value = i32> + SliceByValueGet + SliceByValueSet + SliceByValueRepl,
+        SliceByValueCore<Value = i32> + SliceByValue + SliceByValueMut,
 {
     let r = ..;
     generic_mut(s.index_subslice_mut(r));